BJS2 - 17597 Интеграция с генератором случайных аватаров юзера
+ Исправил по комментариям, навел красоту =)
